refactor(rewards): extract isCampaignIneligible into shared utility
Move the duplicated daysAvailable ineligibility calculation from
OndoCampaignStatsView and OndoCampaignDetailsView into a single
isCampaignIneligible() function in ondoCampaignConstants.ts.
